seminar-03/opeani_research_agent/frontend/requirements.txt
streamlit>=1.30.0
requests>=2.28.2
python-dotenv>=1.0.0
streamlit-extras>=0.3.0 